chore: rename feature blocks components
