+++ apps/marginfi-landing-page-v2/src/components/feature-blocks/index.tsx
@@ -1,2 +1,2 @@
 export * from "./feature-block-centered";
-export * from "./feature-block-grid";
+export * from "./feature-blocks-grid";
